Support for array slices (#265)
* Refactoring, to create separate indexAccessRangeCheck method.
Array slices - WIP.

* Array slices range access returns a triple, to be used for conversion as needed. WIP.

* Refactor array copy so that it enables specification of start-end. WIP

* Initial implementation of array slices. WIP.

* Bug fix for array slices: the length of the slice was computed wrong due to
subtracting of offsets instead of elements. Now, the slice triple
(base, start, end) always contains start and end indices.

* Store end-1 instead of end in array slice triple, in order to utilize existing
checks for bounds.

* - Handled index range access for ArraySlice type.
- Added more checks for the values of start and end index.
TODO: figure out what to do when start == end vs out of bounds.

* Bug fix in range checks for start and end indices.

* Implementation of array slices for byte arrays. Copy from slice to new, allocated memory.

* Store slice end instead of last index in array slice triple (base, start, end).

* Updated execution tests related to array slices, both semantic and syntax tests.

* Updated failing compilation tests.

+++ test/libsolidity/semanticTests/abiEncoderV1/abi_encode_calldata_slice.sol
@@ -36,8 +36,8 @@ contract C {
 	function test_bytes() public {
 		bytes memory test = new bytes(3);
 		test[0] = 0x41; test[1] = 0x42; test[2] = 0x42;
-		for (uint i = 0; i < test.length; i++)
-			for (uint j = i; j <= test.length; j++)
+		for (uint256 i = 0; i < test.length; i++)
+			for (uint256 j = i; j <= test.length; j++)
 			{
 				compare(this.enc_packed_bytes(test, i, j), this.enc_packed_bytes_reference(test, i, j));
 				compare(this.enc_bytes(test, i, j), this.enc_bytes_reference(test, i, j));
@@ -47,8 +47,8 @@ contract C {
 	function test_uint256() public {
 		uint256[] memory test = new uint256[](3);
 		test[0] = 0x41; test[1] = 0x42; test[2] = 0x42;
-		for (uint i = 0; i < test.length; i++)
-			for (uint j = i; j <= test.length; j++)
+		for (uint256 i = 0; i < test.length; i++)
+			for (uint256 j = i; j <= test.length; j++)
 			{
 				compare(this.enc_packed_uint256(test, i, j), this.enc_packed_uint256_reference(test, i, j));
 				compare(this.enc_uint256(test, i, j), this.enc_uint256_reference(test, i, j));
